backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
torch==2.3.0
open_clip_torch==2.24.0
peft==0.10.0
Pillow==10.3.0
numpy==1.26.4
google-generativeai==0.5.4
redis==5.0.4